Handle initial loading state column speculation
